Start refactoring transaction flow [skip ci]
